Downgrading Bootstrap to version 3.3.7 [stable]
Replaced Bootstrap 4 Beta with 3.3.7 Stable since it was causing a lot
of errors.

Created Menu pages for Admin and POCs.

Changed the way the links work so that we have the ability to develop
each individual page in a separate HTML file and have the Content DIV
being able to switch its contents to whatever we choose.

Tested the code works in mobile.